LayoutTests/imported/w3c/web-platform-tests/FileAPI/reading-data-section/filereader_abort-expected.txt


PASS Aborting before read
FAIL Aborting after read assert_equals: Expected abort event, but got error event instead expected "abort" but got "error"
